Iron out some kinks in authentication
It looks like app-state is just going to continue growing -- I'm not
sure that I'm doing the "create user" mutation correctly, but I can't
see a way around this yet without mutating state in the read method
(ew).

Also, checking for a :logged-in key feels grows -- why not check for a
current-user, now that the form state is entirely local?